requirements.txt
tkinter
geopy.geocoders